Nothing important

+++ Engine/World.h
@@ -80,7 +80,7 @@ public:
 		//Functions
 	private:
 		//Functions
-		void AddOres(Block::BlockType type, std::vector<Block>& b, int chanceOfSpawningOnEachBlock);
+		void AddOres(Block::BlockType type, std::vector<Block>& b, float chanceOfSpawningOnEachBlock);
 		void AddOres(Block::BlockType type, std::vector<Block>& b, float chanceOfSpawningCentre, float chanceOfCluster, float minChance, float chanceScalar);
 		void MobSpawning(Mob::MobType type, std::vector<Mob>& m, int propabillity);
 		//Functions
